backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
openai==1.7.1
numpy==1.26.3
python-dotenv==1.0.0
